requirements.txt
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
einops==0.8.0
filelock==3.13.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
imageio==2.36.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
monai==1.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.2.0
portalocker==3.1.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
simsimd==6.2.1
six==1.16.0
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.4.12
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.5.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.5.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.20.1+cu118
torchinfo==1.8.0
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
yacs==0.1.8

